fix: implement V6-compatible PWM calculation and Ackerman fix
- Use direct integer math for PWM calculation (no scaling)
- Apply Ackerman fix to negative angles before error calculation
- Maintain 0-3 degree PWM scaling from lowPWM to highPWM
- Fix Kp data type from float to uint8_t to match PGN format
- Remove excessive PGN 252 debug logging
- Clean up all unnecessary Kp/10 scaling references

The PWM formula now correctly implements: error * Kp + minPWM = PWM output

+++ lib/aio_autosteer/AutosteerProcessor.h
@@ -31,7 +31,7 @@ struct SteerConfig {
 
 // Steer Settings structure (PGN 252)
 struct SteerSettings {
-    float Kp;
+    uint8_t Kp;              // Raw value from AgOpenGPS (actual Kp * 10)
     uint8_t highPWM;
     uint8_t lowPWM;
     uint8_t minPWM;
